% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_lines}
\alias{convert_lines}
\title{Convert lines of RUnit code to testthat code}
\usage{
convert_lines(lines)
}
\arguments{
\item{lines}{An object of class \code{call}.}
}
\value{
An object of class \code{call}.
}
\description{
Converts lines of \code{RUnit} code to the equivalent \code{testthat} code.
}

