\name{convert_switch}
\alias{convert_switch}
\title{Convert a switch block.}
\usage{
convert_switch(x)
}
\arguments{
  \item{x}{A call to the \code{\link[base]{switch}}
  function.}
}
\value{
x the input, with a warning.
}
\description{
Converts a \code{switch} call.  Not yet implemented.
Sorry.
}

