% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_while}
\alias{convert_while}
\title{Convert a while block.}
\usage{
convert_while(x)
}
\arguments{
\item{x}{A call to the \code{\link[base]{while}} function.}
}
\value{
x An object of class \code{call}.
}
\description{
Converts a \code{while} call.
}

