% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert.R
\name{get_filename}
\alias{get_filename}
\title{Get a quoted filename}
\usage{
get_filename(x)
}
\arguments{
\item{x}{A \code{connection} object or a character vector of file paths.}
}
\value{
A character vector of paths to files.
}
\description{
Gets a single quoted filename from a connection or path.
}

