% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_if}
\alias{convert_if}
\title{Convert an if block.}
\usage{
convert_if(x)
}
\arguments{
\item{x}{A call to the \code{\link[base]{if}} function.}
}
\value{
x An object of class \code{call}.
}
\description{
Converts an \code{if} call.
}

